Develop Logout

+++ app/Http/Controllers/BlogController.php
@@ -5,10 +5,22 @@ namespace App\Http\Controllers;
 use App\Models\Blog;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Auth;
 
 class BlogController extends Controller
 {
     //
+
+    private function is_login()
+    {
+        if (Auth::user()) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+
     public function index()
     {
         $blogs = Blog::latest()->paginate(10);
@@ -17,7 +29,11 @@ class BlogController extends Controller
 
     public function create()
     {
-        return view('blog.create');
+        if ($this->is_login()) {
+            return view('blog.create');
+        } else {
+            return redirect('/login');
+        }
     }
     public function store(Request $request)
     {
@@ -48,7 +64,11 @@ class BlogController extends Controller
 
     public function edit(Blog $blog)
     {
-        return view('blog.edit', compact('blog'));
+        if ($this->is_login()) {
+            return view('blog.edit', compact('blog'));
+        } else {
+            return redirect('/login');
+        }
     }
 
     public function update(Request $request, Blog $blog)
@@ -93,16 +113,20 @@ class BlogController extends Controller
     }
     public function destroy($id)
     {
-        $blog = Blog::findOrFail($id);
-        Storage::disk('local')->delete('public/blogs/' . $blog->image);
-        $blog->delete();
+        if ($this->is_login()) {
+            $blog = Blog::findOrFail($id);
+            Storage::disk('local')->delete('public/blogs/' . $blog->image);
+            $blog->delete();
 
-        if ($blog) {
-            //redirect dengan pesan sukses
-            return redirect()->route('blog.index')->with(['success' => 'Data Berhasil Dihapus!']);
+            if ($blog) {
+                //redirect dengan pesan sukses
+                return redirect()->route('blog.index')->with(['success' => 'Data Berhasil Dihapus!']);
+            } else {
+                //redirect dengan pesan error
+                return redirect()->route('blog.index')->with(['error' => 'Data Gagal Dihapus!']);
+            }
         } else {
-            //redirect dengan pesan error
-            return redirect()->route('blog.index')->with(['error' => 'Data Gagal Dihapus!']);
+            return redirect('/login');
         }
     }
 
@@ -113,5 +137,15 @@ class BlogController extends Controller
 
         //render view with blog
         return view('blog.show', compact('blog'));
+    }
+
+    public function showAdmin()
+    {
+        if ($this->is_login()) {
+            $blogs = Blog::latest()->paginate(10);
+            return view('admin', compact('blogs'));
+        } else {
+            return redirect('/login');
+        }
     }
 }
